Full MBC1 implementation (sans its quirks)

+++ gb++/src/Mapper/MBC0/MBC0.cpp
@@ -6,7 +6,7 @@ MBC0::MBC0(uint16_t NumROMBanks, uint16_t NumRAMBanks) : Mapper(NumROMBanks, Num
 MBC0::~MBC0() {
 }
 
-bool MBC0::ROMReadMappedWord(uint16_t address, uint16_t& newAddress) {
+bool MBC0::ROMReadMappedWord(uint16_t address, uint32_t& newAddress) {
 	newAddress = address;
 	return true;
 }
@@ -15,11 +15,11 @@ bool MBC0::ROMWriteMappedWord(uint16_t address, uint8_t data) {
 	return false;
 }
 
-bool MBC0::RAMReadMappedWord(uint16_t address, uint16_t& newAddress) {
+bool MBC0::RAMReadMappedWord(uint16_t address, uint32_t& newAddress) {
 	return false;
 }
 
-bool MBC0::RAMWriteMappedWord(uint16_t address, uint16_t& newAddress) {
+bool MBC0::RAMWriteMappedWord(uint16_t address, uint32_t& newAddress) {
 	return false;
 }
