+++ gb++/src/Mapper/MBC0/MBC0.hpp
@@ -8,11 +8,11 @@ public:
 	MBC0(uint16_t NumROMBanks, uint16_t NumRAMBanks);
 	~MBC0();
 
-	bool ROMReadMappedWord(uint16_t address, uint16_t& newAddress) override;
+	bool ROMReadMappedWord(uint16_t address, uint32_t& newAddress) override;
 	bool ROMWriteMappedWord(uint16_t address, uint8_t data) override;
 
-	bool RAMReadMappedWord(uint16_t address, uint16_t& newAddress) override;
-	bool RAMWriteMappedWord(uint16_t address, uint16_t& newAddress) override;
+	bool RAMReadMappedWord(uint16_t address, uint32_t& newAddress) override;
+	bool RAMWriteMappedWord(uint16_t address, uint32_t& newAddress) override;
 };	
 
 #endif
